Move transactions to their own file
